Add generic GPU frontend and Vulkan conversions
